refactor: rename "layout" to "components"
