+++ src/components/LayoutManager.tsx
