+++ src/components/SplitPane.tsx
