+++ src/components/color-picker/LICENSE
