+++ src/components/color-picker/common/Alpha.jsx
