+++ src/components/color-picker/common/CheckBoard.jsx
