+++ src/components/color-picker/common/EditableInput.jsx
