+++ src/components/color-picker/common/Swatch.jsx
