+++ src/components/color-picker/common/index.js
