+++ src/components/color-picker/common/interaction.jsx
