+++ src/components/color-picker/helpers/alpha.js
