+++ src/components/color-picker/helpers/checkboard.js
