+++ src/components/color-picker/helpers/color.js
