+++ src/components/color-picker/helpers/hue.js
