+++ src/components/color-picker/helpers/saturation.js
