+++ src/components/color-picker/sketch/SketchPresetColors.tsx
