+++ src/components/hooks/useSplitPaneSize.tsx
