+++ stories/layoutManager.stories.tsx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/react';
 import React from 'react';
 
 import { Toolbar } from '../src';
-import LayoutManager, { Layout } from '../src/layout/LayoutManager';
+import LayoutManager, { Layout } from '../src/components/LayoutManager';
 
 export default {
   title: 'Layout/LayoutManager',
